Some Dissertation and Commenting
(~~0~0~~)
Its 19:09 on the 23rd of march 2023 and I am sitting in the abertay whitespace in the university working on the PFNN ue5 project and it was a fun day of working.
Life seems a bit weird currentyly, I think I am a bit confused, maybe cause I am close to graduation, so I am just thinking how the fast the last 4 years were in the long run, the relatively of time is sometimes troubling.
But yeh everyday is me trying to be better then my yesterday but also keeping in mind to enjoy life as much as I can.
That has helped a lot in the way I live my life now.
"
We have two lives and the second one starts when we realize that we only have one
"
But yeh good right now maybe if its not good in the future its just a reminder that things do get better. love ya